Incomplete Lab6 - Maven problems :(
The word editor worked up to a certain point then when I tried adding the last module something with the dependencies went haywire.  With the "Simple Weather Application" I followed the instructions on the website but could not get it to produce the same results.  The build was not able to complete it said.  I have Maven installed, when I used "$mvn -v" I get a valid version and everything so I'm going to have to figure this out over the next couple of weeks I guess.  Screenshots are included.